+++ app/Http/Controllers/API/ApiController.php
@@ -2,17 +2,18 @@
 
 namespace App\Http\Controllers\API;
 
-use App\Models\DetailTransaksi;
-use App\Models\Menu;
-use App\Models\Transaksi;
 use Http;
+use App\Models\Menu;
 use App\Models\Kurir;
 use App\Mail\VerifMail;
 use App\Models\Customer;
+use App\Models\Transaksi;
 use Illuminate\Support\Str;
 use Illuminate\Http\Request;
+use App\Models\DetailTransaksi;
 use App\Http\Controllers\Controller;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\File;
 use Illuminate\Support\Facades\Hash;
 use Illuminate\Support\Facades\Mail;
 use App\Http\Middleware\ApiKeyMiddleware;
@@ -367,6 +368,50 @@ class ApiController extends Controller
             }
             // return $this->sendMassage('status konfirm = 3, status pesanan = 3, status pengiriman = terima', 400, true);
         }
+    }
+
+    public function editCustomer(Request $request)
+    {
+        $token = $request->bearerToken();
+        $user = Customer::where('token', $token)->first();
+
+        if (!$token) {
+            return $this->sendMassage('Tolong masukkan token', 400, false);
+        }
+            $user->nama = $request->input('nama');
+            $user->email = $request->input('email');
+            $user->no_telepon = $request->input('no_telepon');
+            $user->alamat = $request->input('alamat');
+
+            $user->save();
+
+            return $this->sendMassage('Data terupdate', 200, true);
+
+    }
+
+    public function profileImage(Request $request){
+
+        $token = $request->bearerToken();
+        $user = Customer::where('token', $token)->first();
+
+        if(!$token){
+            return $this->sendMassage('Tolong masukkan token', 200, true);
+        }
+
+            if ($request->hasFile('foto')) {
+                $myFile = 'customer/'.$user->foto;
+                if(File::exists($myFile))
+                {
+                    File::delete($myFile);
+                }
+
+                $request->file('foto')->move('customer/', $request->file('foto')->getClientOriginalName());
+                $user->foto=$request->file('foto')->getClientOriginalName();
+
+                $user->save();
+
+                return $this->sendMassage('Foto Profile terupdate', 200, true);
+            }
     }
 
     public function sendMassage($text, $kode, $status)
